test: align hotfix with quality gates
